Refactor integration tests by removing obsolete setup files and consolidating mock configurations
