--- src/integration/setup.js
@@ -1,8 +1,0 @@
-export async function setup() {
-  // Setup für Integrationstests
-  console.log('Integration tests setup started');
-  return () => {
-    // Teardown-Funktion
-    console.log('Integration tests teardown completed');
-  }
-}
