+++ test/integration/test-mocks.js
@@ -1,11 +1,6 @@
 import { vi } from 'vitest'; // vi nur in Testdateien importieren
 
 export function setupApiMocks() {
-  // Mock für dotenv
-  vi.mock('dotenv', () => ({
-    default: { config: vi.fn() },
-    config: vi.fn()
-  }));
 
   // Mock für Externe AI-API
   vi.mock('@google/genai', () => {
